Initial implementation of fetching and applying updates.
It does not work for Windows because Windows does not allow updating in-use executables. I thought it was just executables, but it's also dlls, so I'm giving up on this. It currently hangs on making .new.exe files because xcopy is prompting the user for something.

It works for macOS (unless App Translocation kicks in) and Linux, but honestly it's probably not worth the effort vs. the risk that something will go wrong.